Refactor 02 Forms & Events
I saw a couple of campers seemed to get a bit stuck on the challenge here, so I've broken things up a bit to try and make it a little clearer and easier.

In particular I saw people trying to use `nameFormButton.value` to try and get the value of the input. I think using both `input type="button"` and `input type="text"` might just be adding confusion, and we should use a `<button>` element instead?

Just my thoughts -- I haven't seen as many people run through this than some of the rest of you have!